backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
httpx==0.25.1
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
aioredis==2.0.1
sse-starlette==1.6.5
tenacity==8.2.3